refactor: extract shared TrailHeader component (pixel-perfect)
- Create TrailHeader component with pixel-perfect Figma dimensions:
  145×56 Trilha Ativa box, 82×34 toggle badge, 61×34 fire badge, 76×34 leaf badge
- Toggle supports on/off states (green ON for Trilha, gray OFF for Roleta)
- Edit icon from actual Figma SVG, star icon for leaf badge
- Accepts children prop for extra content (e.g. "Todas as Matérias" on Trilha)
- Remove duplicated header code from Trilha and Roleta screens
